Fix dialyzer warnings and fix bug in jsx_decoder:unescape/5
Thanks to `Dialyzer` and @elbrujohalcon a **bug** was found (and fixed)
in `jsx_decoder:unescape/5`. The bug in question is with a variable (`F`)
in the function's header pattern match that happened to be the same variable
name used inside `?error` macro that is called from this function. The issue
was fixed by renaming `F` to `D` that is actually the character being matched.
A couple of tests were added to test the bug is fixed as well.